lab01

+++ lab01/191220060/sqr.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main()
 {
-    int i,j;
+    long long i,j;
     i=40000;
     j=i*i;
-    printf("The 40000*40000 is %d\n", j);
+    printf("The 40000*40000 is %lld\n", j);
     i=50000;
     j=i*i;
-    printf("The 50000*50000 is %d\n", j);
+    printf("The 50000*50000 is %lld\n", j);
     return 0;
 }
